Compile starter source with tests (#2347)
* Add Gradle task to compile tests against the starter source

* Check compilation of tests + starter source in CI

* Fix stub implementation for `need-for-speed`

* Fix stub implementation for `binary-search`

* Fix stub implementation for `custom-set`

* Fix stub implementation for `perfect-numbers`

* Add missing import for @Ignore in `pov`

* Fix stub implementation for `prime-factors`

* Fix stub implementation for `pythagorean-triplet`

* Fix stub implementation for `rail-fence-cipher`

* Fix stub implementation for `rational-numbers`

* Fix stub implementation for `robot-simulator`

* Fix stub implementation for `roman-numerals`

* Fix stub implementation for `zipper`

+++ exercises/practice/prime-factors/src/main/java/PrimeFactorsCalculator.java
@@ -2,7 +2,7 @@ import java.util.List;
 
 class PrimeFactorsCalculator {
 
-    List<Integer> calculatePrimeFactorsOf(int number) {
+    List<Long> calculatePrimeFactorsOf(long number) {
         throw new UnsupportedOperationException("Delete this statement and write your own implementation.");
     }
